+++ activerecord/lib/active_record/validations/uniqueness.rb
@@ -17,7 +17,13 @@ module ActiveRecord
 
         begin
           relation = build_relation(finder_class, table, attribute, value)
-          relation = relation.and(table[finder_class.primary_key.to_sym].not_eq(record.id)) if record.persisted?
+          if record.persisted?
+            if finder_class.primary_key
+              relation = relation.and(table[finder_class.primary_key.to_sym].not_eq(record.id))
+            else
+              raise UnknownPrimaryKey.new(finder_class, "Can not validate uniqueness for persisted record without primary key.")
+            end
+          end
           relation = scope_relation(record, table, relation)
           relation = finder_class.unscoped.where(relation)
           relation = relation.merge(options[:conditions]) if options[:conditions]